// src/api/models/footerMutate.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * VietProDev CMS Backend
 * Coded by Meu TEAM
 * OpenAPI spec version: 1.0.0
 */
import type { FooterMutateLanguage } from './footerMutateLanguage';
import type { FooterMutateColumnsItem } from './footerMutateColumnsItem';

export interface FooterMutate {
  /** Language of the footer */
  language: FooterMutateLanguage;
  /**
   * Whether the footer is active
   * @nullable
   */
  is_active?: boolean | null;
  /** Columns of the footer */
  columns?: FooterMutateColumnsItem[];
}